Added google/foss flavor to handle wear sync impl
