+++ app/build.gradle
@@ -33,6 +33,17 @@ android {
         }
     }
 
+    flavorDimensions = ["repository"]
+
+    productFlavors {
+        "foss" {
+            dimension "repository"
+        }
+        "google" {
+            dimension "repository"
+        }
+    }
+
     dependenciesInfo {
         // Disables dependency metadata when building APKs.
         includeInApk = false
@@ -66,14 +77,16 @@ android {
 
 dependencies {
     implementation project(path: ':core')
+    implementation project(path: ':sync_manager')
+
     implementation 'androidx.core:core-ktx:1.13.1'
     implementation 'androidx.lifecycle:lifecycle-runtime-ktx:2.8.6'
     implementation 'androidx.activity:activity-compose:1.9.2'
     implementation "androidx.compose.ui:ui:$compose_version"
     implementation "androidx.compose.ui:ui-tooling-preview:$compose_version"
     implementation 'androidx.compose.material3:material3:1.3.0'
-    implementation 'org.jetbrains.kotlinx:kotlinx-coroutines-play-services:1.8.1'
     implementation 'org.jetbrains.kotlinx:kotlinx-coroutines-android:1.8.1'
+
     testImplementation 'junit:junit:4.13.2'
     androidTestImplementation 'androidx.test.ext:junit:1.2.1'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.6.1'
@@ -81,8 +94,6 @@ dependencies {
     debugImplementation "androidx.compose.ui:ui-tooling:$compose_version"
     debugImplementation "androidx.compose.ui:ui-test-manifest:$compose_version"
 
-    //Wear
-    implementation 'com.google.android.gms:play-services-wearable:18.2.0'
 
     // Compose Nav Destinations
     implementation "androidx.navigation:navigation-compose:2.8.1"
