+++ app/src/main/java/audio/omgsoundboard/di/AppModule.kt
@@ -1,11 +1,7 @@
 package audio.omgsoundboard.di
 
 import android.content.Context
-import audio.omgsoundboard.core.data.local.daos.CategoryDao
-import audio.omgsoundboard.core.data.local.daos.SoundsDao
-import audio.omgsoundboard.data.DataLayerRepositoryImpl
 import audio.omgsoundboard.data.SharedPrefRepositoryImpl
-import audio.omgsoundboard.domain.repository.DataLayerRepository
 import audio.omgsoundboard.domain.repository.SharedPrefRepository
 import dagger.Module
 import dagger.Provides
@@ -18,20 +14,9 @@ import javax.inject.Singleton
 @InstallIn(SingletonComponent::class)
 object AppModule {
 
-
     @Singleton
     @Provides
     fun providesStorage(@ApplicationContext context: Context) : SharedPrefRepository {
         return SharedPrefRepositoryImpl(context)
-    }
-
-    @Singleton
-    @Provides
-    fun providesDataLayerRepository(
-        @ApplicationContext context: Context,
-        categoryDao: CategoryDao,
-        soundsDao: SoundsDao,
-    ) : DataLayerRepository {
-        return DataLayerRepositoryImpl(context, categoryDao, soundsDao)
     }
 }
