--- app/src/main/java/audio/omgsoundboard/domain/models/WearNode.kt
@@ -1,7 +1,0 @@
-package audio.omgsoundboard.domain.models
-
-data class WearNode(
-    val id: String = "",
-    val name: String = "",
-    val isNearby: Boolean = false
-)
