--- app/src/main/java/audio/omgsoundboard/domain/repository/DataLayerRepository.kt
@@ -1,10 +1,0 @@
-package audio.omgsoundboard.domain.repository
-
-import audio.omgsoundboard.domain.models.WearNode
-import kotlinx.coroutines.flow.Flow
-
-interface DataLayerRepository {
-    fun getConnectedWearNodesAsFlow(): Flow<List<WearNode>>
-    suspend fun syncDataToWearable(nodeId: String)
-
-}
