+++ app/src/main/java/audio/omgsoundboard/presentation/ui/sounds/SoundsViewModel.kt
@@ -16,13 +16,13 @@ import audio.omgsoundboard.core.domain.repository.PlayerRepository
 import audio.omgsoundboard.core.domain.repository.StorageRepository
 import audio.omgsoundboard.core.utils.Constants.PARTICLES_STATUS
 import audio.omgsoundboard.core.utils.Constants.THEME_TYPE
-import audio.omgsoundboard.domain.repository.DataLayerRepository
 import audio.omgsoundboard.domain.repository.SharedPrefRepository
 import audio.omgsoundboard.presentation.theme.ThemeType
 import audio.omgsoundboard.presentation.theme.toThemeType
 import audio.omgsoundboard.presentation.utils.UiEvent
 import audio.omgsoundboard.presentation.utils.UiText
 import audio.omgsoundboard.presentation.utils.combine
+import audio.omgsoundboard.sync_manager.DataLayerRepository
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.channels.Channel
