+++ build.gradle
@@ -13,4 +13,5 @@ plugins {
     id 'org.jetbrains.kotlin.android' version '2.0.0' apply false
     id 'org.jetbrains.kotlin.plugin.compose' version '2.0.20' apply false
     id 'com.google.devtools.ksp' version '2.0.20-1.0.25' apply false
+    id 'com.android.library' version '8.6.1' apply false
 }
