+++ core/src/main/java/audio/omgsoundboard/core/domain/models/WearNode.kt
@@ -1,0 +1,7 @@
+package audio.omgsoundboard.core.domain.models
+
+data class WearNode(
+    val id: String = "",
+    val name: String = "",
+    val isNearby: Boolean = false
+)
